Extract/requirements.txt
PyYAML>=6.0.1
pandas>=2.0.0
itc-utils  # Version depends on your IBM ServiceFlight package compatibility 